python/requirements.txt
bufbuild-protovalidate-community-nipunn1313-mypy==3.6.0.1.20241127180247+a33202765966
bufbuild-protovalidate-community-nipunn1313-mypy-grpc==3.6.0.1.20241127180247+a33202765966
bufbuild-protovalidate-grpc-python==1.68.0.1.20241127180247+a33202765966
bufbuild-protovalidate-protocolbuffers-pyi==28.3.0.2.20241127180247+a33202765966
bufbuild-protovalidate-protocolbuffers-python==28.3.0.2.20241127180247+a33202765966
certifi==2024.8.30
requests==2.32.3
charset-normalizer==3.4.0
compassiot-api-community-nipunn1313-mypy==3.6.0.1.dev0+cabb2b739966
compassiot-api-community-nipunn1313-mypy-grpc==3.6.0.1.dev0+cabb2b739966
compassiot-api-grpc-python==1.68.0.1.dev0+cabb2b739966
compassiot-api-protocolbuffers-pyi==28.3.0.2.20241129002041+2fc57544816b
compassiot-api-protocolbuffers-python==28.3.0.2.dev0+cabb2b739966
compassiot-model-community-nipunn1313-mypy==3.6.0.1.dev0+ef248955c7dd
compassiot-model-community-nipunn1313-mypy-grpc==3.6.0.1.20241129002041+0dc953f18bd4
compassiot-model-grpc-python==1.68.0.1.20241129002041+0dc953f18bd4
compassiot-model-protocolbuffers-pyi==28.3.0.2.20241129002041+0dc953f18bd4
compassiot-model-protocolbuffers-python==28.3.0.2.20241129002041+0dc953f18bd4
googleapis-googleapis-community-nipunn1313-mypy==3.6.0.1.20241115201650+c0913f24652a
googleapis-googleapis-community-nipunn1313-mypy-grpc==3.6.0.1.20241115201650+c0913f24652a
googleapis-googleapis-grpc-python==1.68.0.1.20241115201650+c0913f24652a
googleapis-googleapis-protocolbuffers-pyi==28.3.0.2.20241115201650+c0913f24652a
googleapis-googleapis-protocolbuffers-python==28.3.0.2.20241115201650+c0913f24652a
grpc-interceptor==0.15.4
grpc-stubs==1.53.0.5
grpcio==1.68.0
grpcio-tools==1.68.0
idna==3.10
protobuf==5.28.3
requests==2.32.3
setuptools==75.6.0
types-protobuf==5.28.3.20241030
urllib3==2.2.3
